integrates division #721 extraction into standings resource
