+++ src/main/groovy/org/domuique/fourteentwo/StandingsResource.groovy
@@ -1,28 +1,14 @@
 package org.domuique.fourteentwo
 
+import java.util.Collection;
+
 class StandingsResource {
 
-    private static final String EXPRESSION = /(\d{3,}?)\s(.*?)\s\d+.*/
-
-    static Collection<Team> get(File file) {
-        file.findAll { (it=~EXPRESSION).matches() }.collect { extract(it) }
-    }
-
-    static def get() {
-        def schedule = downloadDevisionStandings()
-        schedule.eachLine { new Team() }
-        new Team()
-        'fishy!'
-    }
-
-    static Team extract(String line) {
-        def team = (line=~EXPRESSION).collect { match, id, name ->
-           [ 'id': id, 'name': name ]
-        }
-        team.id ? team : null
-    }
-
-    static File downloadDevisionStandings() {
+    private static final String PAGE_BREAK_EXPRESSION = /UPL\/M8 POOL LEAGUE/ 
+    private static final String DIVISION_LINE_EXPRESSION = /Division #(\d{3})/ 
+    private static final String TEAM_STANDING_LINE_EXPRESSION = /(\d{3,}?)\s(.*?)\s\d+.*/
+    
+    static File downloadStandings() {
         def url = new URL('http://www.m8pool.com/pdfs/advsundiv.pdf')
         def file = File.createTempFile('advsundiv', 'pdf')
         def outputStream = file.newOutputStream()
@@ -31,14 +17,35 @@ class StandingsResource {
         file
     }
 
-    static String downloadAndConvertDivisionStandings() {
-        OCR.ocr(StandingsResource.downloadDevisionStandings())
+    static String standings() {
+        OCR.ocr(StandingsResource.downloadStandings())
+    }
+
+    static Collection<String> divide(String standings) {
+        Collection<String> divisions = new ArrayList<String>()
+        String page = ''
+        standings.eachLine {
+            if ( (it=~PAGE_BREAK_EXPRESSION) && !page.empty ) {
+                divisions << page
+                page = ''
+            }
+            if ( !it.empty ) { page += "${it}\n" }
+        }
+        divisions << page
+    }
+
+    static Team extract(String line) {
+        def team = (line=~TEAM_STANDING_LINE_EXPRESSION).collect { match, id, name ->
+           [ 'id': id, 'name': name ]
+        }
+        team.id ? team : null
     }
 
     public static Collection<Team> getTeams() {
         def teams = new ArrayList<Team>()
-        def schedule = downloadAndConvertDivisionStandings()
-        schedule.eachLine { teams << extract(it) }
+        def standings = standings()
+        def divisionStandings = divide(standings).first()
+        divisionStandings.eachLine { teams << extract(it) }
         teams.removeAll([null])
         teams
     }
