+++ src/test/groovy/org/domuique/fourteentwo/StandingsResourceSpec.groovy
@@ -10,12 +10,12 @@ class StandingsResourceSpec extends Specification {
         new File(filepath)
     }
 
-    def 'the team resource can process a Fall 2015 Advanced Sunday schedule test file'() {
+    def 'we can divide standings by division'() {
         when:
-            def file = this.getResourceAsFile('/2015/fall/advsundiv.txt')
-            Collection<Team> teams = StandingsResource.get(file)
+            def standings = this.getResourceAsFile('/2015/fall/advsundiv.txt').text
+            Collection<String> divisions = StandingsResource.divide(standings)
         then:
-            teams?.size() == 36
+            divisions.count { true } == 4
     }
 
     def 'we can extract teams from lines of a schedule file'() {
